新增：随机数，尺子tool；
修改：系统提示词；大部分使用preferenceStore；

修复：滑块弹窗的dismiss逻辑错误；

+++ app/src/main/java/com/hamster/toolbox/ai/Message.kt
@@ -10,7 +10,7 @@ data class Request(
     @SerializedName("stream") val stream: Boolean = false,
     @SerializedName("tools") val tools: List<ToolDefinition>? = null,
     @SerializedName("tool_choice") val toolChoice: String? = null,
-    @SerializedName("temperature") val temperature: Int = 1
+    @SerializedName("temperature") val temperature: Float = 1.0f
 )
 
 @Keep
